pyproject.toml
[project]
name = "tutti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.1",
]
